Move ViewSpy to its own file
