+++ HotelSearch/HotelSearchTests/HotelSearch API/HotelSearchViewModelTests.swift
@@ -136,70 +136,5 @@ class HotelSearchViewModelTests: XCTestCase {
     private func makeHotel(category: String? = nil, description: String? = nil, id: String? = nil, image: URL? = nil, name: String? = nil, stars: Int? = nil) -> Hotel {
         return Hotel(address: nil, amenities: nil, category: category, description: description, gallery: nil, id: id, image: image, isHotel: nil, name: name, price: nil, quantityDescriptors: nil, smallDescription: nil, stars: stars, tags: nil, url: nil)
     }
-    
-    final class ViewSpy: HotelSearchView, HotelSearcher, ImageDataLoader {
-        
-        // MARK: - Hotel Search View
-        
-        enum Message: Hashable {
-            case display(model: [[HotelViewModel]])
-            case display(error: String)
-            case display(loading: Bool)
-            case display(imageData: Data, index: Int, section: Int)
-            case display(imageLoading: Bool, index: Int, section: Int)
-        }
-        
-        private(set) var messages = Set<Message>()
-        
-        func display(_ model: [[HotelViewModel]]) {
-            messages.insert(.display(model: model))
-        }
-        
-        func displayError(_ error: String) {
-            messages.insert(.display(error: error))
-        }
-        
-        func displayLoading(_ isLoading: Bool) {
-            messages.insert(.display(loading: isLoading))
-        }
-        
-        func displayImageData(_ data: Data, for index: Int, section: Int) {
-            messages.insert(.display(imageData: data, index: index, section: section))
-        }
-        
-        func displayImageLoading(_ isLoading: Bool, for index: Int, section: Int) {
-            messages.insert(.display(imageLoading: isLoading, index: index, section: section))
-        }
-        
-        // MARK: - Hotel Searcher
-        
-        private var searchCompletions = [(HotelSearcher.SearchResult) -> Void]()
-        
-        func searchHotel(with searchText: String, completion: @escaping (SearchResult) -> Void) {
-            self.searchCompletions.append(completion)
-        }
-        
-        func completeHotelSearchWith(_ result: HotelSearcher.SearchResult, at index: Int = 0) {
-            self.searchCompletions[index](result)
-        }
-        
-        // MARK: - Image Data Loader
-        
-        private var imageLoadCompletions = [(ImageDataLoader.Result) -> Void]()
-        
-        private struct MockTask: ImageDataLoaderTask {
-            func cancel() { }
-        }
-        
-        func loadImageData(from url: URL, completion: @escaping (ImageDataLoader.Result) -> Void) -> ImageDataLoaderTask {
-            self.imageLoadCompletions.append(completion)
-            return MockTask()
-        }
-        
-        func completeImageLoadWith(_ result: ImageDataLoader.Result, at index: Int = 0) {
-            self.imageLoadCompletions[index](result)
-        }
-        
-    }
 
 }
